Rename project_utils.ts to godot_utils.ts
