+++ src/utils/godot_utils.ts
